pyproject.toml
[tool.poetry]
name = "gsuite-custom-schema-update"
version = "0.1.0"
description = ""
authors = ["Nicolo <nicolo@besharp.it>"]

[tool.poetry.dependencies]
python = "^3.8"
google-api-python-client = "^1.12.5"
google-auth-httplib2 = "^0.0.4"
google-auth-oauthlib = "^0.4.2"
click = "^7.1.2"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
